% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_extract.R
\name{df_extract}
\alias{df_extract}
\title{Title}
\usage{
df_extract(df, which_table)
}
\arguments{
\item{df}{dataframe containing multiple tables}

\item{which_table}{the text from the first cell (1st row, 1st column) of the desired table}
}
\value{
dataframe of the selected contents
}
\description{
Title
}
\examples{
split_tables(data_example("sample_download_file.xlsx"))|> df_extract("Workforce Overview")
}
