% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_gauge.R
\name{gg_gauge}
\alias{gg_gauge}
\title{Build a gauge with red, amber & green background}
\usage{
gg_gauge(pos, title = "", breaks = c(0, 30, 70, 100))
}
\arguments{
\item{pos}{Position of the arrow on the gauge, a value between 1-100}

\item{title}{Title for the individual gauge, defaults to blank}

\item{breaks}{Position of the breaks between colours on the dial}
}
\value{
Returns a ggplot object
}
\description{
Build a gauge with red, amber & green background
}
\examples{
gg_gauge(52)
gg_gauge(65, "How are we performing?")
gg_guage(pos = 75, title = "RAG rating")
}
