% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag_gauges.R
\name{rag_gauges}
\alias{rag_gauges}
\title{DEPO specific function for the 4 dial RAG output}
\usage{
rag_gauges(df)
}
\arguments{
\item{df}{a dataframe with the rag rating as numeric values}
}
\value{
a ggplot object of 4 rag gauges
}
\description{
DEPO specific function for the 4 dial RAG output
}
\examples{
rag_gauges(rag_df) #rag_df is an example dataframe included in this package
rag_gauges(df = rag_df)
}
