% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_row.R
\name{rename_row}
\alias{rename_row}
\title{Title}
\usage{
rename_row(df, row_number)
}
\arguments{
\item{df}{the dataframe with column names in the wrong row}

\item{row_number}{the number of the row containing the desired column names}
}
\value{
a dataframe with the correct column names
}
\description{
Title
}
\examples{
rename_row(df = df, row_number = 2)
}
