% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tables.R
\name{split_tables}
\alias{split_tables}
\title{Split a dataframe read from an untidy sheet in Excel}
\usage{
split_tables(
  filename,
  folder_name_position = "A5",
  skip_rows = 8,
  sheet = "Sheet"
)
}
\arguments{
\item{df}{A single dataframe containing mutliple table with a blank row between them.}
}
\value{
A single dataframe combining all the tables as one, with the columns "folder_name", "table_name", "index_no", "col_names", "value".
}
\description{
Split a dataframe read from an untidy sheet in Excel
}
\examples{
split_tables("sample_download_file.xlsx")
}
